+++ CHANGES.md
@@ -69,6 +69,7 @@
 * Change Trakt view drop down "Show" to reveal Brand-new Shows, Season Premieres, Recommendations and Trending views
 * Change increase number of displayed Trakt shows to 100
 * Add genres and rating to all Trakt shows
+* Add AniDb Random and Hot to Add Show page
 
 [develop changelog]
 Enable Alpha Ratio again now that the secure login page over https is fixed